requirements.txt
kivy>=2.3.0
kivymd>=1.2.0
requests>=2.31.0
Pillow>=10.1.0
python-dotenv>=1.0.0
transformers>=4.45.0
torch>=2.3.0
sentencepiece>=0.2.0
protobuf>=5.0.0
googletrans==4.0.0rc1

